[network] Bring back color coding for WebSocket messages.
The CSS for the color coding was still there and was even updated with
Material 3, but somehow the specificity of the selectors was messed up.
So with this change the sent messages are again highlighted in green.

Drive-by-fix: Also fix the border color for selected, but not focused
rows.

Fixed: 424133329
Screenshot: https://i.imgur.com/BEaHxmm.png
Change-Id: I0217f97b73c090fe0f4588e35f2c6be0ad8d352b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6650629
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>