Add custom eslint rule validate-timing-types
In TypeScript, binary operations on tagged numeric types devolve to the
number type, which we must cast to the desired type. This cast is
unsound, and a mistake can cause bugs that may be difficult to notice
during development.

This CL adds validate-timing-types, a custom eslint rule that leverages
the actual TypeScript types to statically validate that these timing
types are not misused. It uses type casts (`... as Types.Millis`), or
the type of the variable being assigned to, as the expected type - and
checks if the operands of binary expressions match in type.

Also supported is validating some function calls:

- `Array.prototype.reduce` - this would have prevented the bug fixed
  here[1]. Basically, the binary expression inside the reduce callback
  is used to validate against the desired type (whereas tsc may only
  know the return type as "number")
- `Math.min/max` - similar to above, checking the type of each argument
  against the desired type

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6387360

Bug: 406518012
Change-Id: Idb3548665329cbb5dcdfa55a12bc88493a551261
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6395588
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>