Add panel dropdown to bottom menu
This adds the dropdown to open other panels that cannot be shown in
the bottom menu since the window is too narrow. This was the only
exception in Devtools that didn't have that drop down. By adding it,
we make sure that the tabbed pane has the same behavior if the
window is narrow, and facilitates opening other recent panels if
they currently can't be shown.

With the only exception now removed, we can also remove the code
surrounding the option to not add that dropdown.

Fixed: 417901629
Change-Id: I95e32d32aebf4b74c665286c792463bc99ea7605
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6651474
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>