RPP: remove Trace.Extras.Metadata
This CL continues the work to remove SDK deps in models/trace and moves
the metadata calculation into panels/timeline; it didn't need to exist
in models/trace.

I also removed hardware concurrency from the metadata we store as that
setting is no longer in the Performance Panel now so it doesn't really
make a lot of sense to include it.

R=cjamcl@chromium.org

Bug: 376051759
Change-Id: I1b1b7a98dba658ee9fde5f3dbf99a117c3ae530f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6652272
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>