[RPP] Check frame when generating related node map
An easy way to see the problem is to load any trace w/ a LCP in a new
tab (or some other page with lots of elements). The Summary details view
for the LCP element will incorrectly resolve a node element to the
active DOM, just because it found a node with the same backend id.

Instead, check that the event's associated frame is part of the active
target. This catches most issues, though cross-origin navigations may
still be problematic.

Bug: none
Change-Id: I458a8146f6ab5da970194906a439cc8f99157911
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6655579
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>