RPP: fix missing ranges on memory checkbox labels
This CL brings back the ranges that should be shown when you enable the
memory view (which is `CountersGraph` in code) in the performance panel.
The problem was that the devtools-checkbox migration changed the
structure of the DOM and whilst the range was put into the DOM it was
not visible on the page.

This CL updates this by enabling the checkbox to have their user visible
label updated, and ensuring that when we have calculated the range, that
we update it.

R=alexrudenko@chromium.org

Fixed: 427329460
Change-Id: I70b9c9d5c65d6ea4197391bc29e0088918ee8fa3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6662961
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>