Refactor Throttling profiles edit state to be a flexbox
Previously, when trying to add or edit a Throttling profile, it was not possible to scroll the settings items. This caused an issue when devtools were zoomed in because it was not possible to reach to the non-visible items (only with tab).
This changes reafactors the settings component to be a flexbox so it is possible to scroll the settings. Also now the whole title is visible.

Peter lgtm-ed this solution.

Screencast: http://screencast/cast/NTAzMzExMTQ1NTc5MzE1MnwwMGE5ZjFiMC1mNw
Before(one setting not visible): https://screenshot.googleplex.com/4HAhMMjTqC8UgzG

Bug: 403435692
Change-Id: I626556ff1cb21e22a14ededab6fc2f4976a316f8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6662682
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>