[sources] Fix UI eng vision violations in BreakpointsView.ts
This CL migrates the BreakpointsView from a one-off custom element
to a UI.Widget. We re-use the existing render helpers and
consolidate them into a single view function.

The change is rather mechanical but due to the size of the component
is rather verbose. The BreakpointsView is highly interactive so
there a lot of event handlers for the various parts.

We also keep the existing logic that moves the tabfocus around.

One non-trivial change is that the component calculates some strings
for BreakpointItem, notably the description and code snippets for
the label and the tooltip. To keep the diff to a minimum the CL
pre-computes the strings and puts them into a `Map` which is passed
into the view function.

The CL also keeps the shadow root for now: Selectors in
breakpointsView.css are very broad so we'd break existing styling
in DevTools without the shadow root. A separate cleanup needs to
apply naming/nesting to the breakpointsView.css. Afterwards it should
be as easy as removing the shadowDom flag from the `super` call: The
tests already use `contentElement` instead of `shadowRoot`.

R=dsv@chromium.org, kimanh@chromium.org

Fixed: 407751705
Change-Id: Iadc8c45b2d4d2e2f01ae6d92a1ca5dcfa1dd1941
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6667890
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>