AI: add `handleExternalRecordRequest` to TimelinePanel
This CL adds the ability for the TimelinePanel to take an external
request to do a reload & record and return a list of information about
the insights that were found.

The current constraints are that we only return Insights that have the
PerformanceInsightFormatter implementation; this set will grow as we are
going to extend these in Q3.

R=wolfi@chromium.org

Bug: 427883624
Change-Id: I802ea243a84503ce9a495f48111f92849363e330
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6678592
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>