Warm up fonts
I tried multiple preload methods but they did not work.
It seems the best way is just to force the browser to load
font files. This mitigates an issue when for a short period of time
the content does not have the correct font when the font is used
for the first time.

Bug: none
Change-Id: Icf8110be5991bea62f0f82e6568da5fa54ea9f30
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6680411
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>