RPP: collapse insights when sidebar is hidden & restore on open
This CL changes the behaviour of collapsing the sidebar. Now, if you do
it when an insight is active, that insight will be hidden when you
collapse the sidebar. It will then be re-activated when the sidebar
opens.

It's a little janky when it opens, as it animates in just after the
sidebar, but I don't think it's going to be a common case and the effort
to architect this so it doesn't do that is not really warranted, IMO.

R=paulirish@chromium.org

Fixed: 406868070
Change-Id: I226e72ce709b2c12d96b753e3f962775e21b4a3f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6680611
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>