[npm] Replace `npm run watch` with `npm run build -- --watch`.
When `devtools_bundle = false` we can use the fast rebuild for `.ts`
files as well (although this is unsound in case of `const enum`s), but
that was already the case with the `watch_build.js` script.

Also spit out a warning if using `--watch` and `devtools_bundle` isn't
set to `false` in the GN args.

Fixed: 411328609
Change-Id: I9d6f6b51be96b4af581e7c165045e9f0f47e7a0f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6688491
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>