Fix Recorder breakpoint regression
Fixes a regression introduced by https://crrev.com/c/6316671.
Also fixes the test so they fail if the issue is encounter.

What happened was that we changed to Promise.withResolvers which is
an object and one place was not updated to account for this.

The test passes as they check on the same loop/tick so the
assertions passed but if you give them to the next one they failed.
Test were update to account for this and simplified by the use
of a helper.

Fixed: 428651135
Change-Id: Id550fbeae2138b06a3e07007abffaa3ba7f899ae
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6687891
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>