Fix for EsLint jslog list sorting
Our EsLint rule only write to file if we automatically added a new value.
If the value is added manually we only pick up and update the file on the next automatic addition.

To fix the issue of sorting we rely on EsLint linting the front_end/ui/visual_logging/KnownContextValues.ts
In this case we unconditionally sort and write to the file.
This ensures that any manually added values are sorted.

Bug: none
Change-Id: If2561bd2e6b3e7e351d46e836cd44b91920f221d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6695129
Auto-Submit: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>