Store context in AiAgent
The StylingAgent needs to evaluate JS code in the correct execution
context. Before this CL the currently selected element was used both
for determining the execution context, and for evaluating `$0`. This can
be wrong for external requests with cross-proccess iframes, and also if the user selects a different UI element while the conversation is actively running.

Both problems are fixed by keeping the context constant while the
conversation is being executed. Selecting a new element in the UI and
restarting the conversation with a new context still works after this
CL.

Bug: 422019184, 428690606, 419246352
Change-Id: Ica9abdc2e7c165f8b6577817b78cec733620e7a1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6687895
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>