Ensure parent groups of visible messages are always shown in console.
Previously, if a `ConsoleViewMessage` passed the filtering (e.g. error
level), but had a parent group that was collapsed, it would not be added
to the visible view messages (along with its parent group). This
resulted in messages that passed the filtering (and were included in the
counts), but could never actually be shown since the parent group was
not included so there was no way to expand the group and display the
matched message.

The logic has now been adjusted to always show the group (if one
exists), so that users can now expand the group to see their matched
messages. The actual matched message continues to only be added to the
view if there is no parent group or if the parent group is expanded.
This additionally allows the console counter link to properly operate.

Replaces abandoned change on incorrect branch. For more information on the change (e.g. review feedback), see: https://crrev.com/c/6664607

Bug: 40832335
Change-Id: I222eaa18975dee80e4a74536bdeaa3d3ead22fdb
Fixed: 40832335
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6700311
Auto-Submit: Curtis Rodgers <curtisro@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>