Increase timeouts for EsLint rule tests
We need to increase the timeouts for this test because for some test
we need to setup typechecking which sets up TypeScript service to run
inside EsLint. Locally this happens really fast but bots seem to have
a little bit of issue. Increasing to 2x to give them time to set up.
This shouldn't change the time of tests a lot as they are very
stable and we don't experience any flakiness with them.

Bug: none
Change-Id: I42ad8593f318d168864496b5c356b9303bcdb0d9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6700812
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>