Persist network throttling conditions via a storage key
This CL fixes a problem where we rely on the user visible strings when
persisting network throttling conditions to disk, which opens us up to
potential bugs like crbug.com/420384038 where we change the string and
cause DevTools to break for users.

In this CL we change how we persist the network setting to the global
storage. We now persist only a key, and look up the condition when
required.

Along with that, we also persist a list of custom user presets, and each
of those now has a "random" (based on the index) key generated. This
means we persist to disk two settings:

1. A list of the user's custom conditions, including their key, but also
   all their download/upload/etc latency values.
2. The active throttling key, which defaults to `No Throttling`.

When the user changes the value of the key, we look up the condition by
first trying to see if it's predefined (e.g. ships with DevTools) or a
custom one the user has defined. If we fail to find one, we fall back to
`No Throttling`.

This CL is big enough so I didn't want to do any follow-ups, but there's
also some future improvements we can make:

1. Remove `i18nKey` from the conditions; we now don't need it.
2. It looks like `title` might always be a function, we can change the
   type to not allow it to be a `string`.
3. The function that compares network conditions looks at the values, in
   some places we can side-step this and just compare the keys (although
   in other places like the RecorderController we need to compare
   values).

R=szuend@chromium.org

Fixed: 422682525
Change-Id: I2a949a2879e5e303c9aa38993fdad80c159216c4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6694732
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>