[CSSMatchedStyles] Resolve CSS variables inside `CSSKeyframeRule`s
I'm not exactly sure if we need to find the exact CSSStyleDeclaration
where the `@keyframe` rule is used or whether if we can default to just
any style declaration in the element as I expect the resolved variable
for the element to be the same in any case.

Please let me know if I'm overlooking something regarding the
resolution.

Fixed: 364508146
Change-Id: Ic35ce091aac73a2fa9c4aa17fce746be9bc2de06
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6702379
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>