Add a formatter for LLM strings
We are currently working on optimizing the string format for network requests and, as a part of it, we want to get rid of the non-breaking space characters.
This CL adds a formatter for the AI model requests.

Bug: 427939495
Change-Id: I9ac8bd290b0f2a074998592bdfdc0d92f6284a17
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6705235
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Alina Varkki <alinavarkki@chromium.org>