[deps] Add 'source-map-scopes-codec' as a shipped library
This CL pulls in the published tarball of the
'source-map-scopes-codec' package.

We use the standard 'devtools_module' and 'devtools_entrypoint'
to build the library from source.

The CL already adds a dependency on the library via `sdk` so we can
test building the library.

Follow-up CLs will implement the existing scopes decoder with the
library.

R=alexrudenko@chromium.org

Bug: 368222773
Change-Id: I82e27990c01a386bf56fdbc64310d82df3a1016e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6684887
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>