[e2e] Try to restart the browser if it's disconnected
When the browser crashes the underlying Transport (WebSockets) should
close, so we can look at this signal, and try to create a new
browser. For this we need to try to get/check the browser before
each test, using maps to reduce unnecessary check in information.

Also as drive-by clean up some extra issue:
When we try to get screenshot with from closed browser.
When we try to close a closed browser.

Fixed: 430539152
Change-Id: I7c1c3d2037beb6e689d888027303935e5d8b4a4c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6718828
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Nikolay Vitkov <nvitkov@chromium.org>