[network] Add external debug request command
This CL add a command to find a network request with a given URL and
debug it.
It expect the exact url of the request to be found.
If multiple request with the same name are found the first one is
returned.

Bug: 416152806
Change-Id: I2167697185dbe91af5f35288da04f610f02ff3d7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6718061
Auto-Submit: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>