[issues] Render all markdown issue descriptions in unit tests
We had a couple bugs that issues would not render at all due to invalid
markdown issue descriptions. Writing a linter for these is non-trivial
so the next best thing is to straight-up render them all in the
markdown view.

Deprecation issue descriptions are handled separately, as they all use
the same markdown file, and substitute the content in.

For normal issue descriptions we replace placeholders with the
placeholder name, so this doesn't account for any issue that renders
complex things in its place holder.

Some of the issues are currently broken, and need to be fixed on the
backend side.

Drive-by: Fix an issue description that used a broken link.

R=alexrudenko@chromium.org, pfaffe@chromium.org

Fixed: 430801230
Change-Id: I258a69f4977409012bb4f6abafac36877ec73349
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6722231
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>