[ARIAUtils] Implement `status` announcements and move to LiveAnnouncer class
Implements being able to announce some messages as `status`[1] and as
part of this, moves the functionality related to announcements to a
`LiveAnnouncer` class so that we can mock out the `alert` and `status`
methods to use in tests.

Drive-by:
* Update DataGridElement.test.ts to stub `alert` method and do
assertions based on the calls to this method. Otherwise, it depends on
the internal details of the `alert` implementation (it having only one
element, and its text content matching to the provided text)

[1]: b/401064048

Bug: 431767389
Change-Id: Id36da418135791f6db5842fe4f74865160a4e5c8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6734611
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>