Fixes reveal position in RequestResponseView
This broke as a part of a previous CL that changed the view to
align with the ui engineering vision (https://crrev.com/c/6573353).
This CL fixes the bug (binding the correct object to the reveal call).

Fixed: 431007700
Change-Id: I4ff91c358df51db4171b1d0985dded1e00921a0f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6732644
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>