Use compressed network requests format to stringifying multiple network requests
Since the compressed format also requires to provide the format descirprion to the LLM, do not use it for instances when only a single network request is sent. In that case, use the verbose json-like format that has a field name for each value.

Bug: 427939495
Change-Id: I20133bcdcac4c14f0e7bd12d6cb639cfef349385
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6732758
Auto-Submit: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>