Add VE logging to floating buttons
This CL adds visual logging to the floating action buttons in the
Elements, Network, and Sources panels.

A `jslog-context` attribute was added to the `FloatingButton`
component to allow specifying a logging context. This will allow
tracking of clicks on these buttons for analytics.

The visual logging debugging documentation was also updated to
reflect the latest state of the debugging UI.

Fixed: 429381292
Change-Id: I30a3227d99fbd26aa19962a2b56be21cb949ed10
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6728038
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Natallia Harshunova <nharshunova@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>