Fix reliance on VE Chrome feature flag for VE logs in e2e tests
In crrev.com/c/6734500 the VE Logs feature was removed from Chromium,
but we relied on the testing parameter to update the host config for VE
Logs for test.

This CL updates the code for e2e_non_hosted tests to set a setting
`veLogsTestMode` that configures them instead.

In a follow-up, we will remove the VE Logs host config from Chromium &
DevTools entirely, now that we do not rely on the host config, and it is
always enabled, given that the feature flag is now removed and currently
hardcoded to `true` on the backend.

Bug: none
Change-Id: Ib73d9044a32cbdd263891e8f4c24112d293de10b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6764669
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>