Fix Fatal Error getting reported for the next test
There are some issue when dealing with Parallel connection which means
the errors for them are reported close to the page close.

Because of this the DumpError only detects them on the next test.

This CL also fixes an issue with the offending test.

Bug: 422688824
Change-Id: Ie4ab524df907e5c70929ff920e95e209ff3d05db
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6773429
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Liviu Rau <liviurau@chromium.org>