AutoRun: de-duplicate responses in stored output
See the comment in the code for an explanation. This CL ensures that we
do not accidentally duplicate blocks of the response when we write it to
disk. Without this for each query in an example, we log each prior
conversation request/response again.

I also took the opportunity to update the types and fix up the imports.

Bug: 425270170
Change-Id: Id24adf47745354bebf1db088dc5b5c904c495499
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6781163
Reviewed-by: Samiya Caur <samiyac@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>