[ui] Turn 'buildStackTracePreviewContents' into a widget
As part of the 'consistent stack traces' project we need to turn the
JSPresentationUtils into a proper widget to keep track of updates
to the stack trace.

As a first step, lets turn the existing implementation into a widget.
Once the new abstraction is available, we can then also address the
UI vision violations.

Not all use-sites are migrated by this CL but the ones in the network
panel are, to show how this will look like.

R=dsv@chromium.org

Bug: 407752215,433162438
Change-Id: I0a822e4bc620cf34a90e54ccb76f5317c0a0d420
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6796893
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>