[IPP Devtools] Add Filtering of IP Protected Requests to Network Panel
This CL adds in the ability to filter requests by if they went through
IP Protection proxies within the DevTools Network panel. This will allow
for users to see which connections were proxied, and keep track of the
number of requests that were proxied.

The CL also modifies NetworkRequest.ts and NetworkManager.ts to take in
and use the `isIpProtectionUsed` boolean piped in from the back-end.

These changes are gated behind `kIpPrivacyEnableIppInDevTools`,
requiring changes to be made to Runtime.ts for getting that information
for use by the DevTools front-end.

Change-Id: I372b2f9acfbb46a1b6d15e7aac121e83a34f82a7

Bug: 425645896
Change-Id: I372b2f9acfbb46a1b6d15e7aac121e83a34f82a7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6765627
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>