[RPP] Continue updating vertical timestamp marker when mouse down
The vertical timestamp line that aligns the mouse cursor over the
minimap with a position in the flame chart was not updating when the
mouse was held down. This made it difficult to select an exact end
position relative to something visible in the chart.

The fix was to not use mousemove, which doesn't fire during drag, and
instead use pointermove.

Also, capture the pointer during this interaction and hide the overview
info pane until the interaction ends.

Fixed: 393411231
Change-Id: I1218213a706593e06f36726ab7089c0ce86cfd6e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6217297
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Connor Clark <cjamcl@chromium.org>