[bindings] Remove 'isIgnoreListed' from LiveLocation
Now that we have an IgnoreListManager, we can remove 'isIgnoreList'
from the LiveLocation interface. All implementations just retrieve
the UILocation and then pass the UISourceCode to the IgnoreListManager.


With the IgnoreListManager now living in models/workspace, we can
add a bit on the UISourceCode/UILocation that uses the existing
function from the IgnoreLIstMangger.

The major upside of this refactoring is that we cut the number of
'rawLocationToUILocation' calls in half for every linkified
LiveLocation.

R=pfaffe@chromium.org

Bug: 433162438
Change-Id: Ia9376bf44cee9e6c0baa0311f1e76ac79f28b943
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6799328
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>