Fix AI Assistance panel not updating state after move to top/bottom
When moved, panel is being hidden and shown in new location. Event about other panels being revealed and or hidden as a result of the move are sent in between. Ai Assistance panel is unsubscribed from the events while hidden, so it is needed to force update when it is shown again.

Fixed: 434160934
Change-Id: I3521b5e414e4ef944fdb06af3d219930069c14b0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6812254
Auto-Submit: Piotr Paulski <piotrpaulski@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>