Add spinner to AI code completion summary toolbar
Updated spinner to have an `active` attribute.

Added event listener so that we can set the `active` state of spinner
when request is fired to AIDA and response is received.

Both the events in AiCodeCompletion now specify the Panel that triggered
them, this will help us differentiate the ones coming from Console and
Sources.

- Also updated text editor config to handle deletion case better

Bug: 433884684
Change-Id: I0293b2c6af96547867bd7eef45deb26762d02d17
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6780147
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Samiya Caur <samiyac@chromium.org>