RPP: add cache for looking up an entity by a URL
On sites with lots of JS samples and large stacks, we have performance
challenges with the trace engine taking a lot of processing time.

On a relatively small JS trace, I noticed the RendererHandler was taking
~4.1s to parse it, of which a large chunk of time was spent in the
entity mapping code.

I then noticed that when we get an event, we first find its URL, and
then look that URL up in the ThirdPartyWeb dataset. The problem is that
this lookup isn't the cheapest (it's not a straight read from a hashmap
-
https://github.com/patrickhulce/third-party-web/blob/master/lib/create-entity-finder-api.js#L43).
But for this trace in question, I logged 170k instances where we
lookedup a URL that we had already looked up previously.

So, this CL introduces a new cache that maps URL<>Entity. Using this, on
the trace in question, the RendererHandler execution time drops from
~4.1s to ~2.5s, so a nice speed-up which I expect is more prominent on
larger traces of large JS apps where you expect lots of the stack to
have the same URL (e.g. a large React app might have huge stacks where
each event has a URL of `react.min.js`).

R=cjamcl@chromium.org

Bug: 436491188
Change-Id: Ibea925799d9a86d6cb06ee7482b05d36c07d15bf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6819068
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>