[Conversation history] Fix serialization of conversations
Currently, `Conversation.serialize()` leaves the `confirm()` function
of responses of type side-effect unchanged. Consequently, in
`AiHistoryStorage.upsertHistoryEntry()` the `confirm()` function is
persisted in a DevTools setting (only in the setting's in-memory copy,
the copy persisted to storage is stringified which removes the
`confirm()` function).
When the setting is later accessed, `structuredClone()` is executed on
it, which throws when encountering a function.

This bug does not surface in the current implementation of the
`AiAssistancePanel` because the setting is only accessed once, in the
panel's constructor. Current refactoring efforts will change this,
and the conversation serialization needs to be fixed first.

Bug: 427407133
Change-Id: Ib620878245530a5086668a414c85bd70b8be31cf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6827303
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>