RPP: fix extra clone of array in RendererHandler
If you enable the debug experiment, we render samples onto the UI. The
code in `RendererHandler` to do this checked `if(jsSamples)` but
`jsSamples` is an array, so this is always true.

The code within then creates a clone of `allTraceEntries` and then does
another sort on `thread.entries`, both of which we actually only want to
do if the samples _have any length_.

This CL fixes the check to ensure we don't needlessly clone the arrays.

R=alinavarkki@chromium.org

Bug: 436491188
Change-Id: I9fa5e4f0bfaf542d0cf0fd1cdae0dd74825730d4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6827313
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>