[stack_trace] Add scaffolding and some types for the fragment trie
Get the boiler plate out of the way to keep CL size small.

The root node of the trie holds on to it's children via WeakRef.
This allows the GC to cleanup unused stack trace fragments.

In theory, we could use WeakRefs for all child references, to allow
the GC pruning of unused branches. But WeakRefs are not cheap from a
GC perspective so WeakRef-ing the whole Trie would come at a
significant performance cost with questionable benefit.

R=pfaffe@chromium.org

Bug: 433162438
Change-Id: If57cf52301c958ebb3cedbd27bb9fef807fbb5ee
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6837823
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>