Correctly show new badge in Elements panel and subgrid subitems
Before, the new badge wouldn't show up in the Elements panel, as we
passed it in as the wrong argument. Also, for subgrid items we
would show the grid and not the subgrid prompts, as a subgrid is also
a grid.

Fixed: 438129920
Change-Id: I0a5122a6d819360ce8e28c745f2cfc7fbbaed704
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6839780
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>