RPP: disable AnimationFramesHandler by default
We built this when we were going to build out the AnimationFrames track,
which sadly has not happened yet. I don't want to remove the handler,
because we will build the track one day, but we should avoid paying the
cost. This CL disables it by default, but makes it configurable.

On a medium trace with a bunch of JS, this handler takes 50ms, so we
aren't saving a load of time, but we might as well not do the work!

R=alinavarkki@chromium.org

Bug: 436491188
Change-Id: I39883b1753cbb55819595e4970ab8e741a023c44
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6827359
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>