Track ElementsPanel new badge with Promotion Tracker
For the ElementsPanel we currently use two separate features
for showing the new badge: "DevToolsAiDebugWithAi" and the
"DevToolsAiSubmenPrompts" (since DevToolsFreestyler is taken by
the MainView for the 3dot menu).

At launch time, we don't want users that have been in on study group
(e.g. in DevToolsAiSubmenuPrompts) and have seen the new badge, to see
the new badge again if we launch the other study group (e.g. the
DevToolsAiDebugWithAi). Since the back-end doesn't know that it is
the same kind of menu item, additionally register this new badge
with the Promotion Tracker.

Bug: 438371138
Change-Id: I542c51e5a676ed24f234611b151116cba17b7801
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6842541
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>